Tip #5: Template functions
Here's an example of a very simple function that randomly selects an element from a std::vector. A std::vector of what? It doesn't care. Whatever container you pass to it, it will return an element from that container. This might be helpful to you as you build more functionality.