BugFixing
Fixed some strange errors I introduced,  mostly by removing code... I think the main source of the errors was passing an empty vector to the pickrandom function, which called the rolldice function, which used the random.h... I filled it with dummy data and everything was okay. Next step is to change te Character struct to be more of a CharacterSheet struct.